Explorer: the merged search field reaches a plain file's own crumb bar
Extracts the field's box/dropdown/star JSX out of Listing.tsx into a
shared SearchField.tsx, so a folder's listing and a new FileSearchField
(mounted over a plain file in Preview.tsx) render the same component
instead of each carrying a copy. Committing a query over a file, or an
auto-searchable query landing without Enter, navigates to the parent
folder with the query applied as a pushed history entry; the parent's
own listing does the searching. The star gates on the same claimed-bar
mechanism (claimFolderChrome), now claimed with no slot node for a
file since Preview has no split-pane column to relocate the bar into.
